global: add react-overridable and adapt demos
