+++ package.json
@@ -22,6 +22,7 @@
     "qs": "^6.8.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-overridable": "^0.0.2",
     "react-redux": "^7.2",
     "react-scripts": "3.1.1",
     "redux": "^4.0.5",
@@ -44,6 +45,7 @@
     "qs": "^6.8.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-overridable": "^0.0.2",
     "react-redux": "^7.2",
     "react-scripts": "3.2.0",
     "redux": "^4.0.5",
