+++ src/demos/elasticsearch/App.js
@@ -7,7 +7,18 @@
  */
 
 import React, { Component } from 'react';
-import { Container, Grid, Menu, Label } from 'semantic-ui-react';
+import {
+  Button,
+  Container,
+  Grid,
+  Menu,
+  Label,
+  Card,
+  Image,
+  Item,
+} from 'semantic-ui-react';
+import _truncate from 'lodash/truncate';
+import { OverridableContext } from 'react-overridable';
 import {
   BucketAggregation,
   ReactSearchKit,
@@ -44,7 +55,7 @@ const customAggComp = (title, containerCmp) => {
   ) : null;
 };
 
-const customAggValuesContainerCmp = valuesCmp => (
+const customAggValuesContainerCmp = (valuesCmp) => (
   <Menu.Menu>{valuesCmp}</Menu.Menu>
 );
 
@@ -69,53 +80,122 @@ const customAggValueCmp = (
   );
 };
 
+class Tags extends Component {
+  onClick = (event, value) => {
+    window.history.push({
+      search: `${window.location.search}&f=tags_agg:${value}`,
+    });
+    event.preventDefault();
+  };
+
+  render() {
+    return this.props.tags.map((tag, index) => (
+      <Button
+        key={index}
+        size="mini"
+        onClick={(event) => this.onClick(event, tag)}
+      >
+        {tag}
+      </Button>
+    ));
+  }
+}
+
+const ElasticSearchResultsListItem = ({ result, index }) => {
+  return (
+    <Item key={index} href={`#`}>
+      <Item.Image
+        size="small"
+        src={result.picture || 'http://placehold.it/200'}
+      />
+      <Item.Content>
+        <Item.Header>
+          {result.first_name} {result.last_name}
+        </Item.Header>
+        <Item.Description>
+          {_truncate(result.about, { length: 200 })}
+        </Item.Description>
+        <Item.Extra>
+          <Tags tags={result.tags} />
+        </Item.Extra>
+      </Item.Content>
+    </Item>
+  );
+};
+
+const ElasticSearchResultsGridItem = ({ result, index }) => {
+  return (
+    <Card fluid key={index} href={`#`}>
+      <Image src={result.picture || 'http://placehold.it/200'} />
+      <Card.Content>
+        <Card.Header>
+          {result.first_name} {result.last_name}
+        </Card.Header>
+        <Card.Description>
+          {_truncate(result.about, { length: 200 })}
+        </Card.Description>
+      </Card.Content>
+      <Card.Content extra>
+        <Tags tags={result.tags} />
+      </Card.Content>
+    </Card>
+  );
+};
+
+const overriddenComponents = {
+  'ResultsList.item.elasticsearch': ElasticSearchResultsListItem,
+  'ResultsGrid.item.elasticsearch': ElasticSearchResultsGridItem,
+};
+
 export class App extends Component {
   render() {
     return (
-      <ReactSearchKit searchApi={searchApi}>
-        <Container>
-          <Grid>
-            <Grid.Row>
-              <Grid.Column width={3} />
-              <Grid.Column width={10}>
-                <SearchBar />
-              </Grid.Column>
-              <Grid.Column width={3} />
-            </Grid.Row>
-          </Grid>
-          <Grid relaxed style={{ padding: '2em 0' }}>
-            <Grid.Row columns={2}>
-              <Grid.Column width={4}>
-                <BucketAggregation
-                  title="Tags"
-                  agg={{ field: 'tags', aggName: 'tags_agg' }}
-                  renderElement={customAggComp}
-                  renderValuesContainerElement={customAggValuesContainerCmp}
-                  renderValueElement={customAggValueCmp}
-                />
-                <BucketAggregation
-                  title="Employee Types"
-                  agg={{
-                    field: 'employee_type.type',
-                    aggName: 'type_agg',
-                    childAgg: {
-                      field: 'employee_type.subtype',
-                      aggName: 'subtype_agg',
-                    },
-                  }}
-                />
-              </Grid.Column>
-              <Grid.Column width={12}>
-                <ResultsLoader>
-                  <EmptyResults />
-                  <Error />
-                  <OnResults />
-                </ResultsLoader>
-              </Grid.Column>
-            </Grid.Row>
-          </Grid>
-        </Container>
-      </ReactSearchKit>
+      <OverridableContext.Provider value={overriddenComponents}>
+        <ReactSearchKit searchApi={searchApi}>
+          <Container>
+            <Grid>
+              <Grid.Row>
+                <Grid.Column width={3} />
+                <Grid.Column width={10}>
+                  <SearchBar />
+                </Grid.Column>
+                <Grid.Column width={3} />
+              </Grid.Row>
+            </Grid>
+            <Grid relaxed style={{ padding: '2em 0' }}>
+              <Grid.Row columns={2}>
+                <Grid.Column width={4}>
+                  <BucketAggregation
+                    title="Tags"
+                    agg={{ field: 'tags', aggName: 'tags_agg' }}
+                    renderElement={customAggComp}
+                    renderValuesContainerElement={customAggValuesContainerCmp}
+                    renderValueElement={customAggValueCmp}
+                  />
+                  <BucketAggregation
+                    title="Employee Types"
+                    agg={{
+                      field: 'employee_type.type',
+                      aggName: 'type_agg',
+                      childAgg: {
+                        field: 'employee_type.subtype',
+                        aggName: 'subtype_agg',
+                      },
+                    }}
+                  />
+                </Grid.Column>
+                <Grid.Column width={12}>
+                  <ResultsLoader>
+                    <EmptyResults />
+                    <Error />
+                    <OnResults />
+                  </ResultsLoader>
+                </Grid.Column>
+              </Grid.Row>
+            </Grid>
+          </Container>
+        </ReactSearchKit>
+      </OverridableContext.Provider>
     );
   }
 }
