+++ src/lib/components/ActiveFilters/ActiveFilters.js
@@ -9,7 +9,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Label, Icon } from 'semantic-ui-react';
-import { Overridable } from 'react-overridable';
+import Overridable from 'react-overridable';
 import { buildUID } from '../../util';
 
 export default class ActiveFilters extends Component {
