+++ src/lib/components/AutocompleteSearchBar/AutocompleteSearchBar.js
@@ -10,7 +10,7 @@ import _ from 'lodash';
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Input } from 'semantic-ui-react';
-import { Overridable } from 'react-overridable';
+import Overridable from 'react-overridable';
 import './AutocompleteSearchBar.scss';
 import { buildUID } from '../../util';
