+++ src/lib/components/BucketAggregation/BucketAggregationValues.js
@@ -9,7 +9,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Checkbox, List } from 'semantic-ui-react';
-import { Overridable } from 'react-overridable';
+import Overridable from 'react-overridable';
 import { buildUID } from '../../util';
 
 export default class BucketAggregationValues extends Component {
